rename file to lowercase
