+++ credsweeper/filters/group/__init__.py
@@ -8,5 +8,5 @@ from credsweeper.filters.group.pem_pattern import PEMPattern
 from credsweeper.filters.group.structured_token import StructuredToken
 from credsweeper.filters.group.token_pattern import TokenPattern
 from credsweeper.filters.group.token_base32 import TokenBase32
-from credsweeper.filters.group.JWT import JWT
+from credsweeper.filters.group.jwt import JWT
 from credsweeper.filters.group.url_credentials_group import UrlCredentialsGroup
